Release constraints in to Public Domain; remove latest constraint
This changes the licensing status of the constraint imports to be under the
Public Domain using The Unlicense. This allows anyone to use, consume, include,
distribute, change, and destroy the source code as they deem appropriate. The
only thing that shall remain under the `MIT License` are the files in the
`scripts/` directory.

Signed-off-by: Tim Heckman <t@heckman.io>